cleanup and doc
