+++ Assets/Prefabs/Lobby.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3aed21dda64ed49b3ad5fa62007a45a2
+folderAsset: yes
+timeCreated: 1517441797
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
