+++ Assets/Scripts/Lobby.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3a1a19fc1f9b44e209b7f1c6929acb13
+folderAsset: yes
+timeCreated: 1517441748
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
